Revert "Cretae seperate class for models"
