--- Message.swift
@@ -1,22 +1,0 @@
-//
-//  Message.swift
-//  NepTalk
-//
-//  Created by Kshitiz Bista on 2022-01-01.
-//
-
-import Foundation
-import MessageKit
-
-struct Message: MessageType {
-    var sender: SenderType
-    var messageId: String
-    var sentDate: Date
-    var kind: MessageKind
-}
-
-struct Sender: SenderType {
-    var senderId: String
-    var displayName: String
-    var photoURL: String
-}
