+++ NepTalk/Controllers/ConversationsViewController.swift
@@ -8,6 +8,20 @@
 import UIKit
 import JGProgressHUD
 
+struct Conversation {
+    let id: String
+    let receiverName: String
+    let receiverEmail: String
+    let receiverUID: String
+    let latestMessage: LatestMessage
+}
+
+struct LatestMessage {
+    let date: String
+    let message: String
+    let isRead: Bool
+}
+
 class ConversationsViewController: UIViewController {
     
     private var conversations = [Conversation]()
