+++ NepTalk/Resources/DatabaseManager.swift
@@ -15,6 +15,12 @@ final class DatabaseManager {
     static let shared = DatabaseManager()
     private let database = Database.database().reference()
     
+    static func safeEmail(email: String) -> String {
+        var safeEmail = email.replacingOccurrences(of: ".", with: "-")
+        safeEmail = safeEmail.replacingOccurrences(of: "@", with: "-")
+        return safeEmail
+    }
+    
     public enum DatabaseError: Error {
         case failedToFetch
     }
@@ -422,5 +428,22 @@ extension DatabaseManager {
             }
             completion(true)
         }
+    }
+}
+
+struct AppUser {
+    let uid: String
+    let firstName: String
+    let lastName: String
+    var email: String
+    var profilePictureFileName: String {
+        return "\(uid)_profile_pic.png"
+    }
+    
+    init(uid: String, firstName: String, lastName: String, email: String) {
+        self.uid = uid
+        self.firstName = firstName
+        self.lastName = lastName
+        self.email = email
     }
 }
